+++ src/main/java/com/iexec/core/detector/replicate/ReplicateResultUploadTimeoutDetector.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 IEXEC BLOCKCHAIN TECH
+ * Copyright 2020-2024 IEXEC BLOCKCHAIN TECH
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,6 +16,7 @@
 
 package com.iexec.core.detector.replicate;
 
+import com.iexec.common.replicate.ReplicateStatusUpdate;
 import com.iexec.core.detector.Detector;
 import com.iexec.core.replicate.Replicate;
 import com.iexec.core.replicate.ReplicatesService;
@@ -84,8 +85,8 @@ public class ReplicateResultUploadTimeoutDetector implements Detector {
             log.info("detected replicate with resultUploadTimeout [chainTaskId:{}, replicate:{}, currentStatus:{}]",
                     chainTaskId, uploadingReplicate.getWalletAddress(), uploadingReplicate.getCurrentStatus());
 
-            replicatesService.updateReplicateStatus(chainTaskId, uploadingReplicate.getWalletAddress(),
-                    RESULT_UPLOAD_FAILED);
+            replicatesService.updateReplicateStatus(
+                    chainTaskId, uploadingReplicate.getWalletAddress(), ReplicateStatusUpdate.poolManagerRequest(RESULT_UPLOAD_FAILED));
             taskUpdateRequestManager.publishRequest(task.getChainTaskId());
         }
     }