+++ src/main/java/com/iexec/core/task/listener/ReplicateListeners.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 IEXEC BLOCKCHAIN TECH
+ * Copyright 2020-2024 IEXEC BLOCKCHAIN TECH
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -28,6 +28,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.context.event.EventListener;
 import org.springframework.stereotype.Component;
 
+import static com.iexec.common.replicate.ReplicateStatus.FAILED;
 import static com.iexec.common.replicate.ReplicateStatusCause.TASK_NOT_ACTIVE;
 
 @Slf4j
@@ -86,7 +87,7 @@ public class ReplicateListeners {
          * */
         if (ReplicateStatus.getUncompletableStatuses().contains(newStatus)) {
             replicatesService.updateReplicateStatus(event.getChainTaskId(),
-                    event.getWalletAddress(), ReplicateStatus.FAILED);
+                    event.getWalletAddress(), ReplicateStatusUpdate.poolManagerRequest(FAILED));
         }
 
         /*